Updates & Bug fixes
Returning pointer from function was not easy as it required ponter variable to be static.
Static variables are shared by all function calls but those pointers dont need to be shared.
Functions returning ponter were updated to dynamically allocate memory for ponters.

Dynamic allocated memory is not local to function which would allow returning poiter without being static.
Function for slicing array does not work as expected but works from strings even if not always.
There seems to be some runtime issues for string but ordinary arrays are never sliced.

Utility functions are working but not always as expected especially those involving slicing.
There is a bug but cannot be detected due to lack of deep knowledge about C language.